refactor: function and variable name
