+++ lockfile/audit/package.json
@@ -30,7 +30,6 @@
   },
   "homepage": "https://github.com/pnpm/pnpm/blob/main/lockfile/audit#readme",
   "devDependencies": {
-    "@pnpm/reviewing.dependencies-hierarchy": "workspace:*",
     "@pnpm/audit": "workspace:*",
     "@pnpm/constants": "workspace:*",
     "@pnpm/lockfile-file": "workspace:*",
