+++ reviewing/dependencies-hierarchy/src/PackageNode.ts
@@ -1,15 +1,1 @@
-export interface PackageNode {
-  alias: string
-  circular?: true
-  dependencies?: PackageNode[]
-  dev?: boolean
-  isPeer: boolean
-  isSkipped: boolean
-  isMissing: boolean
-  name: string
-  optional?: true
-  path: string
-  resolved?: string
-  searched?: true
-  version: string
-}
+export type { PackageNode } from '@pnpm/types'
